phan quyen (auth)

+++ resources/views/articles/show.blade.php
@@ -2,14 +2,19 @@
 @extends('layout')
 @section('content')
     <h1>{{ $article->title }}</h1>
+    <label>Publish at: {{ $article->published_at }}</label>
     <hr/>
     <article>
         <div class="body">{{ $article->body }}</div>
     </article><br>
-    {!! link_to(action('ArticlesController@edit', [$article->id]), 'Edit', ['class' => 'btn btn-primary']) !!}
 
-    <br>
-
-    {!! delete_form(['articles', $article->id]) !!} 
+    {{-- Và lúc show bài viết cũng tương tự với nút Edit và Delete --}}
+    @if (Auth::check())
+        <br/>
+        {!! link_to(route('articles.edit', [$article->id]), 'Edit', ['class' => 'btn btn-primary']) !!}
+        <br/>
+        <br/>
+        {!! delete_form(['articles', $article->id]) !!}
+    @endif
 
 @endsection
